ML4T MC3 P2 everything working
Need to decide on what to use for training data. Since using moving averages, what is the best way to fill in gaps, i.e., beginning of time. When I fill it backwards, in sample is awesome, out of sample less so. When I remove back filled data, both in and out of sample perform ok. Perhaps this would be something to discuss in the report.